--- a/src/maxminddb/geoip2.rs
+++ b/src/maxminddb/geoip2.rs
@@ -1,40 +1,42 @@
 /// GeoIP2 Country record
 #[derive(Deserialize, Serialize, Clone, Debug)]
-pub struct Country {
-    pub continent: Option<model::Continent>,
-    pub country: Option<model::Country>,
-    pub registered_country: Option<model::Country>,
-    pub represented_country: Option<model::RepresentedCountry>,
+pub struct Country<'a> {
+    #[serde(borrow)]
+    pub continent: Option<model::Continent<'a>>,
+    pub country: Option<model::Country<'a>>,
+    pub registered_country: Option<model::Country<'a>>,
+    pub represented_country: Option<model::RepresentedCountry<'a>>,
     pub traits: Option<model::Traits>,
 }
 
 /// GeoIP2 City record
 #[derive(Deserialize, Serialize, Clone, Debug)]
-pub struct City {
-    pub city: Option<model::City>,
-    pub continent: Option<model::Continent>,
-    pub country: Option<model::Country>,
-    pub location: Option<model::Location>,
-    pub postal: Option<model::Postal>,
-    pub registered_country: Option<model::Country>,
-    pub represented_country: Option<model::RepresentedCountry>,
-    pub subdivisions: Option<Vec<model::Subdivision>>,
+pub struct City<'a> {
+    pub city: Option<model::City<'a>>,
+    #[serde(borrow)]
+    pub continent: Option<model::Continent<'a>>,
+    pub country: Option<model::Country<'a>>,
+    pub location: Option<model::Location<'a>>,
+    pub postal: Option<model::Postal<'a>>,
+    pub registered_country: Option<model::Country<'a>>,
+    pub represented_country: Option<model::RepresentedCountry<'a>>,
+    pub subdivisions: Option<Vec<model::Subdivision<'a>>>,
     pub traits: Option<model::Traits>,
 }
 
 /// GeoIP2 ISP record
 #[derive(Deserialize, Serialize, Clone, Debug)]
-pub struct Isp {
+pub struct Isp<'a> {
     pub autonomous_system_number: Option<u32>,
-    pub autonomous_system_organization: Option<String>,
-    pub isp: Option<String>,
-    pub organization: Option<String>,
+    pub autonomous_system_organization: Option<&'a str>,
+    pub isp: Option<&'a str>,
+    pub organization: Option<&'a str>,
 }
 
 /// GeoIP2 Connection-Type record
 #[derive(Deserialize, Serialize, Clone, Debug)]
-pub struct ConnectionType {
-    pub connection_type: Option<String>,
+pub struct ConnectionType<'a> {
+    pub connection_type: Option<&'a str>,
 }
 
 /// GeoIP2 Anonymous Ip record
@@ -56,66 +58,67 @@ pub struct DensityIncome {
 
 /// GeoIP2 Domain record
 #[derive(Deserialize, Serialize, Clone, Debug)]
-pub struct Domain {
-    pub domain: Option<String>,
+pub struct Domain<'a> {
+    pub domain: Option<&'a str>,
 }
 
 /// GeoIP2 Asn record
 #[derive(Deserialize, Serialize, Clone, Debug)]
-pub struct Asn {
+pub struct Asn<'a> {
     pub autonomous_system_number: Option<u32>,
-    pub autonomous_system_organization: Option<String>,
+    pub autonomous_system_organization: Option<&'a str>,
 }
 
 pub mod model {
     use std::collections::BTreeMap;
 
     #[derive(Deserialize, Serialize, Clone, Debug)]
-    pub struct City {
+    pub struct City<'a> {
         pub geoname_id: Option<u32>,
-        pub names: Option<BTreeMap<String, String>>,
+        #[serde(borrow)]
+        pub names: Option<BTreeMap<&'a str, &'a str>>,
     }
 
     #[derive(Deserialize, Serialize, Clone, Debug)]
-    pub struct Continent {
-        pub code: Option<String>,
+    pub struct Continent<'a> {
+        pub code: Option<&'a str>,
         pub geoname_id: Option<u32>,
-        pub names: Option<BTreeMap<String, String>>,
+        pub names: Option<BTreeMap<&'a str, &'a str>>,
     }
 
     #[derive(Deserialize, Serialize, Clone, Debug)]
-    pub struct Country {
+    pub struct Country<'a> {
         pub geoname_id: Option<u32>,
         pub is_in_european_union: Option<bool>,
-        pub iso_code: Option<String>,
-        pub names: Option<BTreeMap<String, String>>,
+        pub iso_code: Option<&'a str>,
+        pub names: Option<BTreeMap<&'a str, &'a str>>,
     }
 
     #[derive(Deserialize, Serialize, Clone, Debug)]
-    pub struct Location {
+    pub struct Location<'a> {
         pub latitude: Option<f64>,
         pub longitude: Option<f64>,
         pub metro_code: Option<u16>,
-        pub time_zone: Option<String>,
+        pub time_zone: Option<&'a str>,
     }
 
     #[derive(Deserialize, Serialize, Clone, Debug)]
-    pub struct Postal {
-        pub code: Option<String>,
+    pub struct Postal<'a> {
+        pub code: Option<&'a str>,
     }
 
     #[derive(Deserialize, Serialize, Clone, Debug)]
-    pub struct RepresentedCountry {
+    pub struct RepresentedCountry<'a> {
         pub geoname_id: Option<u32>,
-        pub iso_code: Option<String>,
-        pub names: Option<BTreeMap<String, String>>, // pub type: Option<String>,
+        pub iso_code: Option<&'a str>,
+        pub names: Option<BTreeMap<&'a str, &'a str>>, // pub type: Option<&'a str>,
     }
 
     #[derive(Deserialize, Serialize, Clone, Debug)]
-    pub struct Subdivision {
+    pub struct Subdivision<'a> {
         pub geoname_id: Option<u32>,
-        pub iso_code: Option<String>,
-        pub names: Option<BTreeMap<String, String>>,
+        pub iso_code: Option<&'a str>,
+        pub names: Option<BTreeMap<&'a str, &'a str>>,
     }
 
     #[derive(Deserialize, Serialize, Clone, Debug)]