This PR adds performance enhancements to maxminddb-rust.  It is broken into three commits.

![2020-06-06-225255_484x543_scrot](https://user-images.githubusercontent.com/2483388/83961516-b4a44e80-a848-11ea-82d4-8efbf430c3fa.png)

Checking out upstream/master, `cargo bench` gives me the following:

```
     Running target/release/deps/lookup-8016720bc456e2d4
maxminddb               time:   [2.1721 ms 2.1730 ms 2.1741 ms]
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high severe

maxminddb_par           time:   [190.76 us 192.28 us 193.60 us]
```

## 1. Reduce internal allocations

This commit mostly modifies DataRecord to use &str instead of String, and comes with some decent performance improvements:

```
     Running target/release/deps/lookup-8016720bc456e2d4
maxminddb               time:   [1.9167 ms 1.9180 ms 1.9208 ms]
                        change: [-11.884% -11.651% -11.283%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 10 measurements (20.00%)
  1 (10.00%) high mild
  1 (10.00%) high severe

maxminddb_par           time:   [163.56 us 165.95 us 168.37 us]
                        change: [-15.047% -13.843% -12.652%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 10 measurements (20.00%)
  1 (10.00%) low mild
  1 (10.00%) high mild
```

## 2. Add opt-in unsafe string decode feature

This commit adds a new opt-in feature that skips utf8 validation, an intensive operation that may not be needed in performance-critical environments where the maxminddb is trusted.  When benched with the feature turned on (vs the previous commit):

```
maxminddb               time:   [1.8062 ms 1.8103 ms 1.8175 ms]
                        change: [-5.9584% -5.4895% -5.0301%] (p = 0.00 < 0.05)
                        Performance has improved.

maxminddb_par           time:   [146.40 us 147.32 us 148.30 us]
                        change: [-11.260% -9.8868% -8.3574%] (p = 0.00 < 0.05)
                        Performance has improved.
```

## 3. Use &str in returned types

This commit changes the API completely.  All Strings in the geoip structs are now returned as references.  However, the results are great if ownership is not needed (vs the previous commit and with the unsafe-str-decode feature turned on):

```
maxminddb               time:   [1.2839 ms 1.3063 ms 1.3352 ms]
                        change: [-29.340% -28.380% -27.321%] (p = 0.00 < 0.05)
                        Performance has improved.

maxminddb_par           time:   [119.04 us 120.71 us 122.97 us]
                        change: [-19.156% -16.821% -14.017%] (p = 0.00 < 0.05)
                        Performance has improved.
```

While the benchmarks can vary, a project I'm working on, which uses this as a backend, goes from handling 200,000 requests per second to over 250,000 (25%).