@oschwald 2020-06-07
Nice work! These changes look great. Thanks for the detailed summary.

With these changes, the performance approximately matches my [Go reader](https://github.com/oschwald/maxminddb-golang). I'd assume there are improvements that can be made in the decoding of tree records and control bytes that would help further. Unfortunately, I haven't spent much time improving this library. The Go one I use at work, making it more likely to see improvements.